Added User models and its functionalities, CRUD operations and authentication using JWT

+++ backend/controllers/goalController.js
@@ -1,10 +1,11 @@
 const asyncHandler = require('express-async-handler')
 const Goal = require('../models/goalModel')
+const User = require('../models/userModel')
 //@desc Get Goals
 //@route GET /api/goals
 //@access PRIVATE
 const getGoals = asyncHandler(async(req, res)=> {
-    const goals = await Goal.find()
+    const goals = await Goal.find({user: req.user.id})
     res.json(goals)
 })
 
@@ -19,7 +20,8 @@ const setGoal = asyncHandler(async(req, res)=> {
         throw new Error("Please add a text field")
     }
     const goal = await Goal.create({
-        text: req.body.text
+        text: req.body.text,
+        user: req.user.id,
     })
     res.status(200).json(goal)
     console.log(req.body.text)
@@ -38,6 +40,17 @@ const updateGoal = asyncHandler(async(req, res)=> {
         throw new Error('Goal not found')
     }
 
+    const user = await User.findById(req.user.id)
+    if(!user){
+        res.status(401)
+        throw new Error('User not found')
+    }
+
+    if(goal.user.toString() !== user.id){
+        res.status(401)
+        throw new Error('User not authorized')
+    }
+
     const updatedGoal = await Goal.findByIdAndUpdate(req.params.id, req.body, {new: true})
 
     res.json(updateGoal)
@@ -54,6 +67,16 @@ const deleteGoal = asyncHandler(async(req, res)=> {
         res.status(404)
         throw new Error('Goal not found')
 
+    }
+    const user = await User.findById(req.user.id)
+    if(!user){
+        res.status(401)
+        throw new Error('User not found')
+    }
+
+    if(goal.user.toString() !== user.id){
+        res.status(401)
+        throw new Error('User not authorized')
     }
     const deletedGoal = await Goal.findByIdAndDelete(req.params.id)
     res.json(deletedGoal)
